Removing PayloadUtils.sol
