+++ contracts/Tesseract.sol
@@ -11,11 +11,11 @@ import "@openzeppelin/contracts-upgradeable/access/OwnableUpgradeable.sol";
 
 import "./interfaces/IWormholeBridge.sol";
 import "./interfaces/ITesseract.sol";
-import "./utils/PayloadUtils.sol";
+import "./utils/Versionable.sol";
 
 //import "hardhat/console.sol";
 
-contract Tesseract is ITesseract, Initializable, PayloadUtils, OwnableUpgradeable, UUPSUpgradeable {
+contract Tesseract is ITesseract, Versionable, Initializable, OwnableUpgradeable, UUPSUpgradeable {
   using AddressUpgradeable for address;
   using SafeMathUpgradeable for uint256;
