--- contracts/interfaces/IPayloadUtils.sol
@@ -1,37 +1,0 @@
-// SPDX-License-Identifier: MIT
-pragma solidity 0.8.11;
-
-// Author: Francesco Sullo <francesco@sullo.co>
-// (c) 2022+ SuperPower Labs Inc.
-
-interface IPayloadUtils {
-  function version() external pure returns (uint256);
-
-  function validateInput(
-    uint256 tokenType,
-    uint256 lockupTime,
-    uint256 tokenAmountOrID
-  ) external pure returns (bool);
-
-  function deserializeInput(uint256 payload)
-    external
-    pure
-    returns (
-      uint256 tokenType,
-      uint256 lockupTime,
-      uint256 tokenAmountOrID
-    );
-
-  function deserializeDeposit(uint256 payload)
-    external
-    pure
-    returns (
-      uint256 tokenType,
-      uint256 lockedFrom,
-      uint256 lockedUntil,
-      uint256 mainIndex,
-      uint256 tokenAmountOrID
-    );
-
-  function getIndexFromPayload(uint256 payload) external pure returns (uint256);
-}
