+++ test/MainPool.test.js
@@ -1,6 +1,6 @@
 const {expect, assert, use} = require("chai");
 
-const {serializeInput} = require("../scripts/lib/PayloadUtils");
+const {serializeInput, fromMainDepositToTransferPayload} = require("../scripts/lib/PayloadUtils");
 
 const {
   initEthers,
@@ -228,5 +228,68 @@ describe("#MainPool", function () {
       await mainPool.withdrawSSynr(0, treasury.address);
       expect(await sSynr.balanceOf(treasury.address)).equal(amount);
     });
+  });
+
+  describe("#serializeInput", async function () {
+    beforeEach(async function () {
+      await initAndDeploy();
+    });
+
+    it("should serialize input", async function () {
+      const amount = ethers.utils.parseEther("10000");
+
+      const payload = await serializeInput(
+        SYNR_STAKE, // SYNR
+        365, // 1 year
+        amount
+      );
+
+      expect(payload).equal("1000000000000000000000036502");
+    });
+
+    it("should throw invalid token", async function () {
+      const amount = ethers.utils.parseEther("10000");
+
+      expect(serializeInput(120, 365, amount)).revertedWith("PayloadUtils: invalid token type");
+    });
+
+    it("should throw not a mobland pass", async function () {
+      const amount = ethers.utils.parseEther("10000");
+      serializeInput;
+      expect(serializeInput(2, 365, amount)).revertedWith("PayloadUtils: Not a Mobland SYNR Pass token ID");
+    });
+
+    it("should throw amount of range", async function () {
+      const amount = ethers.utils.parseEther("1000000000000");
+
+      expect(serializeInput(1, 365, amount)).revertedWith("PayloadUtils: tokenAmountOrID out of range");
+    });
+
+    it("should throw lockedTime out of range", async function () {
+      const amount = ethers.utils.parseEther("10000");
+
+      expect(serializeInput(1, 1e5, amount)).revertedWith("PayloadUtils: lockedTime out of range");
+    });
+  });
+
+  describe("#deserializeInput", async function () {
+    beforeEach(async function () {
+      await initAndDeploy();
+    });
+
+    it("should deserialize", async function () {
+      const amount = ethers.utils.parseEther("10000");
+
+      const payload = await serializeInput(
+        SYNR_PASS_STAKE_FOR_BOOST, // SYNR
+        365, // 1 year
+        200
+      );
+      const deserialize = await mainPool.deserializeInput(payload);
+
+      expect(parseInt(deserialize)).equal(SYNR_PASS_STAKE_FOR_BOOST, 365, amount);
+    });
+
+    // TODO add a fake payload and verify if it fails
   });
 });
